out.ltsreg {BMA}	R Documentation

out.ltsreg

Description

Function to identify potential outliers

Usage

out.ltsreg(x, y, delta)

Arguments

x
the design matrix

y
observations

delta
the threshold set by the user. Standardized residuals from least trimmed squares regression that are larger than delta are identified as potential outliers

Value

A 0/1 vector indicating whether each observation is a potential outlier. The function was designed for use with the variable and outlier selection function MC3.REG

Author(s)

Jennifer A. Hoeting

See Also

MC3.REG

[Package BMA version 3.18.21 Index]